Fix wordlist e2e tests
This commit fixes the routing for all of the wordlist e2e testing so they all pass with the updated paths. More specifically, they were changed to work with context packs. We also made some of the tests more robust. This commit also adds additional context packs into the json so we have a wider range of context packs.

Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>
Co-authored-by:  Elena Lam <lam00071@morris.umn.edu>