Fix broken buttons, include main word in word forms and wordlist colors
There were some buttons that were not going to the right activities and are now fixed. Another thing that was addressed in this commit was KK's feedback to include the main word in the word forms. Wordlists now show the words in them according to color as well. Addresses issue #44.

Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>
Co-authored-by:  Jacob Jenness <jenne077@morris.umn.edu>
Co-authored-by:  Elena Lam <lam00071@morris.umn.edu>
Co-authored-by:  Biruk Mengistu <mengi024@morris.umn.edu>