Achieve 100% test coverage in server
This commit adds the necessary code to increase branch coverage of the methods in Word River Controller. This commit also adds a new test that checks that wordlists with a null name or a zero character name do not validate. This commit also adds a test that ensure that a NotFoundResponse is thrown when getWordLists is called on a context pack that has a wordlist value of null. 

Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>
Co-authored-by:  Jacob Jenness <jenne077@morris.umn.edu>
Co-authored-by:  Elena Lam <lam00071@morris.umn.edu>
Co-authored-by:  Biruk Mengistu <mengi024@morris.umn.edu>