Clean up code
This is a rather large commit, but it goes through most of the files that we've worked on to get rid of unnecessary comments, `console.log` statements that we used for testing, and unused import statements. We also tweaked a couple of places in the code that did redundant or unneeded things, like defining unused variables. 
Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>
Co-authored-by:  Jacob Jenness <jenne077@morris.umn.edu>
Co-authored-by:  Elena Lam <lam00071@morris.umn.edu>
Co-authored-by:  Biruk Mengistu <mengi024@morris.umn.edu>