Fix README and Replace Reference to Team Chicken in index.html
This commit modifies the links in the README file in order to point things toward our team's repository for this iteration and not the previous one. Additionally, in index.html, a reference to Team Chicken was replaced with Vibing Cats. This was updated to reflect the name of our team.

Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>
Co-authored-by:  Jacob Jenness <jenne077@morris.umn.edu>
Co-authored-by:  Elena Lam <lam00071@morris.umn.edu>
Co-authored-by:  Biruk Mengistu <mengi024@morris.umn.edu>