Fix failing Karma tests
This commit fixes various failing Karma tests. Most of the failures had to do with a design change of the url of the home page. It was changed to /home, so we had to change the karma tests to reflect that design change. There were also a couple of tests where  there were no providers for the login service. There was also one test where the provided path was not correct. Issue #34

Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by:  Kyle Day <day00090@morris.umn.edu>
Co-authored-by:  Jacob Jenness <jenne077@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>
Co-authored-by:  Elena Lam <lam00071@morris.umn.edu>
Co-authored-by:  Biruk Mengistu <mengi024@morris.umn.edu>