Fix failing Cypress tests
This commit fixes various failing Cypress tests. Most of these failures were due to the fact that during the tests we weren't logged into website. A couple of the failures were also due to the fact that cypress would try to get elements of a page before the page had time to load. Issue #33

Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by:  Kyle Day <day00090@morris.umn.edu>
Co-authored-by:  Jacob Jenness <jenne077@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>
Co-authored-by:  Elena Lam <lam00071@morris.umn.edu>
Co-authored-by:  Biruk Mengistu <mengi024@morris.umn.edu>