Update learner-data-related code to get learner data response
This commit includes minor changes to the structure of LearnerData.java and LearnerDataController.java so that Unity can get a JSON response that comes from the Mongo database. Something was broken about how the parsing was happening before, and possibly also something was wrong with initial values for word counts.

Co-authored-by: Ben Burgess <burge337@morris.umn.edu>