Implements Post request/API for LearnerData and updates getLearnerData()
The post request either creates a new document for the collection if there wasn't already one in there, or it updates an existing document.

getLearnerData() now creates a document with only the learnerId (and object id that mongo generates for us) if the database query didn't find anything but the learnerId passed wasn't null. (as is the case for all new learners) The check on learnerId should eventually be more useful/strict, as invalid learnerIds can still be passed right now.

Serves as server code for commit: 028f39edce566531a4136c8e420448a8f91f6314
in kidstech/storybuilder repository