Get spec tests for sentenceControllerSpec to pass
This commit gets the tests for the sentenceControllerSpec to pass. The problem was that the constructor for the tests did not have the correct number of arguments. So, we just added null, since we do not have any tests related to the story controller aspect of sentence controller. This is a temporary fix.

Co-authored-by: Natasha-Zebrev <zebre001@morris.umn.edu>