Fix story times to improve filtering
Story Builder uses the time the most recent story was saved to decide which sentences to display in the story building view. For new users, there was no saved story from which to get that time. This update makes it so that a default time is used in that case. It also makes it so that translation between 12 hour and 24 hour clocks works (which was failing previously).